@@ -8,9 +8,10 @@ import (
 	"testing"
 	"time"
 
-	"github.com/mvisonneau/gitlab-ci-pipelines-exporter/pkg/config"
 	"github.com/stretchr/testify/assert"
 	"github.com/urfave/cli/v2"
+
+	"github.com/mvisonneau/gitlab-ci-pipelines-exporter/pkg/config"
 )
 
 func NewTestContext() (ctx *cli.Context, flags *flag.FlagSet) {
@@ -28,10 +29,14 @@ func NewTestContext() (ctx *cli.Context, flags *flag.FlagSet) {
 }
 
 func TestConfigure(t *testing.T) {
-	var cfg config.Config
-	var err error
+	var (
+		cfg config.Config
+		err error
+	)
+
 	f, err := ioutil.TempFile(".", "test-*.yml")
 	assert.NoError(t, err)
+
 	defer os.Remove(f.Name())
 
 	// Webhook endpoint enabled
@@ -44,17 +49,20 @@ func TestConfigure(t *testing.T) {
 
 	// Undefined gitlab-token
 	flags.String("gitlab-token", "", "")
+
 	_, err = configure(ctx)
 	assert.Error(t, err)
 
 	// Valid configuration
 	flags.Set("gitlab-token", "secret")
+
 	cfg, err = configure(ctx)
 	assert.NoError(t, err)
 	assert.Equal(t, "secret", cfg.Gitlab.Token)
 
 	// Invalid config file syntax
 	ioutil.WriteFile(f.Name(), []byte("["), 0o644)
+
 	cfg, err = configure(ctx)
 	assert.Error(t, err)
 
@@ -72,9 +80,19 @@ server:
 
 	// Defining the webhook secret token
 	flags.String("webhook-secret-token", "secret", "")
+
 	cfg, err = configure(ctx)
 	assert.NoError(t, err)
 	assert.Equal(t, "secret", cfg.Server.Webhook.SecretToken)
+
+	// Test health url flag
+	healthURL := "https://gitlab.com/-/readiness?token"
+	flags.String("gitlab-health-url", healthURL, "")
+
+	cfg, err = configure(ctx)
+	assert.NoError(t, err)
+	assert.Equal(t, cfg.Gitlab.HealthURL, healthURL)
+	assert.True(t, cfg.Gitlab.EnableHealthCheck)
 }
 
 func TestExit(t *testing.T) {
