@@ -13,6 +13,8 @@ func TestNew(t *testing.T) {
 	c.Log.Level = "info"
 	c.Log.Format = "text"
 
+	c.OpenTelemetry.GRPCEndpoint = ""
+
 	c.Server.ListenAddress = ":8080"
 	c.Server.Metrics.Enabled = true
 
@@ -21,6 +23,8 @@ func TestNew(t *testing.T) {
 	c.Gitlab.EnableHealthCheck = true
 	c.Gitlab.EnableTLSVerify = true
 	c.Gitlab.MaximumRequestsPerSecond = 1
+	c.Gitlab.BurstableRequestsPerSecond = 5
+	c.Gitlab.MaximumJobsQueueSize = 1000
 
 	c.Pull.ProjectsFromWildcards.OnInit = true
 	c.Pull.ProjectsFromWildcards.Scheduled = true
@@ -56,7 +60,7 @@ func TestNew(t *testing.T) {
 	c.ProjectDefaults.Pull.Environments.ExcludeStopped = true
 
 	c.ProjectDefaults.Pull.Refs.Branches.Enabled = true
-	c.ProjectDefaults.Pull.Refs.Branches.Regexp = `^main|master$`
+	c.ProjectDefaults.Pull.Refs.Branches.Regexp = `^(?:main|master)$`
 	c.ProjectDefaults.Pull.Refs.Branches.ExcludeDeleted = true
 
 	c.ProjectDefaults.Pull.Refs.Tags.Enabled = true
