@@ -4,7 +4,7 @@ import (
 	"net/url"
 )
 
-// Global is used for globally shared exporter config
+// Global is used for globally shared exporter config.
 type Global struct {
 	// InternalMonitoringListenerAddress can be used to access
 	// some metrics related to the exporter internals
