@@ -4,10 +4,10 @@ import (
 	"github.com/creasty/defaults"
 )
 
-// Wildcard is a specific handler to dynamically search projects
+// Wildcard is a specific handler to dynamically search projects.
 type Wildcard struct {
 	// ProjectParameters holds parameters specific to the projects which
-	// will be discovered using this wildcard
+	// will be discovered using this wildcard.
 	ProjectParameters `yaml:",inline"`
 
 	Search   string        `yaml:"search"`
@@ -25,8 +25,9 @@ type WildcardOwner struct {
 // Wildcards ..
 type Wildcards []Wildcard
 
-// NewWildcard returns a new wildcard with the default parameters
+// NewWildcard returns a new wildcard with the default parameters.
 func NewWildcard() (w Wildcard) {
 	defaults.MustSet(&w)
+
 	return
 }
