@@ -1,29 +1,29 @@
 package controller
 
 import (
-	"context"
 	"fmt"
 	"net/http"
 	"testing"
 
+	"github.com/stretchr/testify/assert"
+
 	"github.com/mvisonneau/gitlab-ci-pipelines-exporter/pkg/config"
 	"github.com/mvisonneau/gitlab-ci-pipelines-exporter/pkg/schemas"
-	"github.com/stretchr/testify/assert"
 )
 
 func TestPullProjectsFromWildcard(t *testing.T) {
-	c, mux, srv := newTestController(config.Config{})
+	ctx, c, mux, srv := newTestController(config.Config{})
 	defer srv.Close()
 
 	mux.HandleFunc("/api/v4/projects",
 		func(w http.ResponseWriter, r *http.Request) {
-			fmt.Fprint(w, `[{"id":1,"path_with_namespace":"foo","jobs_enabled":false},{"id":2,"path_with_namespace":"bar","jobs_enabled":true}]`)
+			fmt.Fprint(w, `[{"id":2,"path_with_namespace":"bar","jobs_enabled":true}]`)
 		})
 
 	w := config.NewWildcard()
-	assert.NoError(t, c.PullProjectsFromWildcard(context.Background(), w))
+	assert.NoError(t, c.PullProjectsFromWildcard(ctx, w))
 
-	projects, _ := c.Store.Projects()
+	projects, _ := c.Store.Projects(ctx)
 	p1 := schemas.NewProject("bar")
 
 	expectedProjects := schemas.Projects{
