@@ -1,18 +1,18 @@
 package controller
 
 import (
-	"context"
 	"fmt"
 	"net/http"
 	"testing"
 
+	"github.com/stretchr/testify/assert"
+
 	"github.com/mvisonneau/gitlab-ci-pipelines-exporter/pkg/config"
 	"github.com/mvisonneau/gitlab-ci-pipelines-exporter/pkg/schemas"
-	"github.com/stretchr/testify/assert"
 )
 
 func TestGetRefs(t *testing.T) {
-	c, mux, srv := newTestController(config.Config{})
+	ctx, c, mux, srv := newTestController(config.Config{})
 	defer srv.Close()
 
 	mux.HandleFunc("/api/v4/projects/foo/repository/branches",
@@ -35,7 +35,7 @@ func TestGetRefs(t *testing.T) {
 	p.Pull.Refs.Tags.Regexp = `^v`
 	p.Pull.Refs.MergeRequests.Enabled = true
 
-	foundRefs, err := c.GetRefs(p)
+	foundRefs, err := c.GetRefs(ctx, p)
 	assert.NoError(t, err)
 
 	ref1 := schemas.NewRef(p, schemas.RefKindBranch, "main")
@@ -50,7 +50,7 @@ func TestGetRefs(t *testing.T) {
 }
 
 func TestPullRefsFromProject(t *testing.T) {
-	c, mux, srv := newTestController(config.Config{})
+	ctx, c, mux, srv := newTestController(config.Config{})
 	defer srv.Close()
 
 	mux.HandleFunc("/api/v4/projects/foo",
@@ -69,13 +69,13 @@ func TestPullRefsFromProject(t *testing.T) {
 		})
 
 	p1 := schemas.NewProject("foo")
-	assert.NoError(t, c.PullRefsFromProject(context.Background(), p1))
+	assert.NoError(t, c.PullRefsFromProject(ctx, p1))
 
 	ref1 := schemas.NewRef(p1, schemas.RefKindBranch, "main")
 	expectedRefs := schemas.Refs{
 		ref1.Key(): ref1,
 	}
 
-	projectsRefs, _ := c.Store.Refs()
+	projectsRefs, _ := c.Store.Refs(ctx)
 	assert.Equal(t, expectedRefs, projectsRefs)
 }
