@@ -1,9 +1,12 @@
 package controller
 
 import (
+	"context"
+
+	log "github.com/sirupsen/logrus"
+
 	"github.com/mvisonneau/gitlab-ci-pipelines-exporter/pkg/schemas"
 	"github.com/mvisonneau/gitlab-ci-pipelines-exporter/pkg/store"
-	log "github.com/sirupsen/logrus"
 )
 
 func metricLogFields(m schemas.Metric) log.Fields {
@@ -13,32 +16,29 @@ func metricLogFields(m schemas.Metric) log.Fields {
 	}
 }
 
-func storeGetMetric(s store.Store, m *schemas.Metric) {
-	if err := s.GetMetric(m); err != nil {
-		log.WithFields(
-			metricLogFields(*m),
-		).WithField(
-			"error", err.Error(),
-		).Errorf("reading metric from the store")
+func storeGetMetric(ctx context.Context, s store.Store, m *schemas.Metric) {
+	if err := s.GetMetric(ctx, m); err != nil {
+		log.WithContext(ctx).
+			WithFields(metricLogFields(*m)).
+			WithError(err).
+			Errorf("reading metric from the store")
 	}
 }
 
-func storeSetMetric(s store.Store, m schemas.Metric) {
-	if err := s.SetMetric(m); err != nil {
-		log.WithFields(
-			metricLogFields(m),
-		).WithField(
-			"error", err.Error(),
-		).Errorf("writing metric in the store")
+func storeSetMetric(ctx context.Context, s store.Store, m schemas.Metric) {
+	if err := s.SetMetric(ctx, m); err != nil {
+		log.WithContext(ctx).
+			WithFields(metricLogFields(m)).
+			WithError(err).
+			Errorf("writing metric from the store")
 	}
 }
 
-func storeDelMetric(s store.Store, m schemas.Metric) {
-	if err := s.DelMetric(m.Key()); err != nil {
-		log.WithFields(
-			metricLogFields(m),
-		).WithField(
-			"error", err.Error(),
-		).Errorf("deleting metric from the store")
+func storeDelMetric(ctx context.Context, s store.Store, m schemas.Metric) {
+	if err := s.DelMetric(ctx, m.Key()); err != nil {
+		log.WithContext(ctx).
+			WithFields(metricLogFields(m)).
+			WithError(err).
+			Errorf("deleting metric from the store")
 	}
 }
