@@ -7,15 +7,18 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/mvisonneau/gitlab-ci-pipelines-exporter/pkg/config"
-	"github.com/mvisonneau/gitlab-ci-pipelines-exporter/pkg/schemas"
 	log "github.com/sirupsen/logrus"
 	goGitlab "github.com/xanzy/go-gitlab"
+
+	"github.com/mvisonneau/gitlab-ci-pipelines-exporter/pkg/config"
+	"github.com/mvisonneau/gitlab-ci-pipelines-exporter/pkg/schemas"
 )
 
-func (c *Controller) processPipelineEvent(e goGitlab.PipelineEvent) {
-	var refKind schemas.RefKind
-	refName := e.ObjectAttributes.Ref
+func (c *Controller) processPipelineEvent(ctx context.Context, e goGitlab.PipelineEvent) {
+	var (
+		refKind schemas.RefKind
+		refName = e.ObjectAttributes.Ref
+	)
 
 	// TODO: Perhaps it would be nice to match upon the regexp to validate
 	// that it is actually a merge request ref
@@ -28,23 +31,132 @@ func (c *Controller) processPipelineEvent(e goGitlab.PipelineEvent) {
 		refKind = schemas.RefKindBranch
 	}
 
-	c.triggerRefMetricsPull(schemas.NewRef(
+	c.triggerRefMetricsPull(ctx, schemas.NewRef(
 		schemas.NewProject(e.Project.PathWithNamespace),
 		refKind,
 		refName,
 	))
 }
 
-func (c *Controller) triggerRefMetricsPull(ref schemas.Ref) {
+func (c *Controller) processJobEvent(ctx context.Context, e goGitlab.JobEvent) {
+	var (
+		refKind schemas.RefKind
+		refName = e.Ref
+	)
+
+	if e.Tag {
+		refKind = schemas.RefKindTag
+	} else {
+		refKind = schemas.RefKindBranch
+	}
+
+	project, _, err := c.Gitlab.Projects.GetProject(e.ProjectID, nil)
+	if err != nil {
+		log.WithContext(ctx).
+			WithError(err).
+			Error("reading project from GitLab")
+
+		return
+	}
+
+	c.triggerRefMetricsPull(ctx, schemas.NewRef(
+		schemas.NewProject(project.PathWithNamespace),
+		refKind,
+		refName,
+	))
+}
+
+func (c *Controller) processPushEvent(ctx context.Context, e goGitlab.PushEvent) {
+	if e.CheckoutSHA == "" {
+		var (
+			refKind = schemas.RefKindBranch
+			refName string
+		)
+
+		// branch refs in push events have "refs/heads/" prefix
+		if branch, found := strings.CutPrefix(e.Ref, "refs/heads/"); found {
+			refName = branch
+		} else {
+			log.WithContext(ctx).
+				WithFields(log.Fields{
+					"project-name": e.Project.Name,
+					"ref":          e.Ref,
+				}).
+				Error("extracting branch name from ref")
+
+			return
+		}
+
+		_ = deleteRef(ctx, c.Store, schemas.NewRef(
+			schemas.NewProject(e.Project.PathWithNamespace),
+			refKind,
+			refName,
+		), "received branch deletion push event from webhook")
+	}
+}
+
+func (c *Controller) processTagEvent(ctx context.Context, e goGitlab.TagEvent) {
+	if e.CheckoutSHA == "" {
+		var (
+			refKind = schemas.RefKindTag
+			refName string
+		)
+
+		// tags refs in tag events have "refs/tags/" prefix
+		if tag, found := strings.CutPrefix(e.Ref, "refs/tags/"); found {
+			refName = tag
+		} else {
+			log.WithContext(ctx).
+				WithFields(log.Fields{
+					"project-name": e.Project.Name,
+					"ref":          e.Ref,
+				}).
+				Error("extracting tag name from ref")
+
+			return
+		}
+
+		_ = deleteRef(ctx, c.Store, schemas.NewRef(
+			schemas.NewProject(e.Project.PathWithNamespace),
+			refKind,
+			refName,
+		), "received tag deletion tag event from webhook")
+	}
+}
+
+func (c *Controller) processMergeEvent(ctx context.Context, e goGitlab.MergeEvent) {
+	ref := schemas.NewRef(
+		schemas.NewProject(e.Project.PathWithNamespace),
+		schemas.RefKindMergeRequest,
+		strconv.Itoa(e.ObjectAttributes.IID),
+	)
+
+	switch e.ObjectAttributes.Action {
+	case "close":
+		_ = deleteRef(ctx, c.Store, ref, "received merge request close event from webhook")
+	case "merge":
+		_ = deleteRef(ctx, c.Store, ref, "received merge request merge event from webhook")
+	default:
+		log.
+			WithField("merge-request-event-type", e.ObjectAttributes.Action).
+			Debug("received a non supported merge-request event type as a webhook")
+	}
+}
+
+func (c *Controller) triggerRefMetricsPull(ctx context.Context, ref schemas.Ref) {
 	logFields := log.Fields{
 		"project-name": ref.Project.Name,
 		"ref":          ref.Name,
 		"ref-kind":     ref.Kind,
 	}
 
-	refExists, err := c.Store.RefExists(ref.Key())
+	refExists, err := c.Store.RefExists(ctx, ref.Key())
 	if err != nil {
-		log.WithFields(logFields).WithError(err).Error("reading ref from the store")
+		log.WithContext(ctx).
+			WithFields(logFields).
+			WithError(err).
+			Error("reading ref from the store")
+
 		return
 	}
 
@@ -52,58 +164,81 @@ func (c *Controller) triggerRefMetricsPull(ref schemas.Ref) {
 	if !refExists {
 		p := schemas.NewProject(ref.Project.Name)
 
-		projectExists, err := c.Store.ProjectExists(p.Key())
+		projectExists, err := c.Store.ProjectExists(ctx, p.Key())
 		if err != nil {
-			log.WithFields(logFields).WithError(err).Error("reading project from the store")
+			log.WithContext(ctx).
+				WithFields(logFields).
+				WithError(err).
+				Error("reading project from the store")
+
 			return
 		}
 
 		// Perhaps the project is discoverable through a wildcard
 		if !projectExists && len(c.Config.Wildcards) > 0 {
-			for id, w := range c.Config.Wildcards {
+			for _, w := range c.Config.Wildcards {
 				// If in all our wildcards we have one which can potentially match the project ref
-				// received, we trigger a scan
+				// received, we trigger a pull of the project
 				matches, err := isRefMatchingWilcard(w, ref)
 				if err != nil {
-					log.WithError(err).Warn("checking if the ref matches the wildcard config")
+					log.WithContext(ctx).
+						WithError(err).
+						Warn("checking if the ref matches the wildcard config")
+
 					continue
 				}
 
 				if matches {
-					c.ScheduleTask(context.TODO(), schemas.TaskTypePullProjectsFromWildcard, strconv.Itoa(id), strconv.Itoa(id), w)
-					log.WithFields(logFields).Info("project ref not currently exported but its configuration matches a wildcard, triggering a pull of the projects from this wildcard")
+					c.ScheduleTask(context.TODO(), schemas.TaskTypePullProject, ref.Project.Name, ref.Project.Name, w.Pull)
+					log.WithFields(logFields).Info("project ref not currently exported but its configuration matches a wildcard, triggering a pull of the project")
 				} else {
 					log.WithFields(logFields).Debug("project ref not matching wildcard, skipping..")
 				}
 			}
+
 			log.WithFields(logFields).Info("done looking up for wildcards matching the project ref")
+
 			return
 		}
 
 		if projectExists {
 			// If the project exists, we check that the ref matches it's configuration
-			if err := c.Store.GetProject(&p); err != nil {
-				log.WithFields(logFields).WithError(err).Error("reading project from the store")
+			if err := c.Store.GetProject(ctx, &p); err != nil {
+				log.WithContext(ctx).
+					WithFields(logFields).
+					WithError(err).
+					Error("reading project from the store")
+
 				return
 			}
 
 			matches, err := isRefMatchingProjectPullRefs(p.Pull.Refs, ref)
 			if err != nil {
-				log.WithError(err).Error("checking if the ref matches the project config")
+				log.WithContext(ctx).
+					WithError(err).
+					Error("checking if the ref matches the project config")
+
 				return
 			}
 
 			if matches {
 				ref.Project = p
-				if err = c.Store.SetRef(ref); err != nil {
-					log.WithFields(logFields).WithError(err).Error("writing ref in the store")
+
+				if err = c.Store.SetRef(ctx, ref); err != nil {
+					log.WithContext(ctx).
+						WithFields(logFields).
+						WithError(err).
+						Error("writing ref in the store")
+
 					return
 				}
+
 				goto schedulePull
 			}
 		}
 
 		log.WithFields(logFields).Info("ref not configured in the exporter, ignoring pipeline webhook")
+
 		return
 	}
 
@@ -114,91 +249,123 @@ schedulePull:
 	c.ScheduleTask(context.TODO(), schemas.TaskTypePullRefMetrics, string(ref.Key()), ref)
 }
 
-func (c *Controller) processDeploymentEvent(e goGitlab.DeploymentEvent) {
-	c.triggerEnvironmentMetricsPull(schemas.Environment{
-		ProjectName: e.Project.PathWithNamespace,
-		Name:        e.Environment,
-	})
+func (c *Controller) processDeploymentEvent(ctx context.Context, e goGitlab.DeploymentEvent) {
+	c.triggerEnvironmentMetricsPull(
+		ctx,
+		schemas.Environment{
+			ProjectName: e.Project.PathWithNamespace,
+			Name:        e.Environment,
+		},
+	)
 }
 
-func (c *Controller) triggerEnvironmentMetricsPull(env schemas.Environment) {
+func (c *Controller) triggerEnvironmentMetricsPull(ctx context.Context, env schemas.Environment) {
 	logFields := log.Fields{
 		"project-name":     env.ProjectName,
 		"environment-name": env.Name,
 	}
 
-	envExists, err := c.Store.EnvironmentExists(env.Key())
+	envExists, err := c.Store.EnvironmentExists(ctx, env.Key())
 	if err != nil {
-		log.WithFields(logFields).WithError(err).Error("reading environment from the store")
+		log.WithContext(ctx).
+			WithFields(logFields).
+			WithError(err).
+			Error("reading environment from the store")
+
 		return
 	}
 
 	if !envExists {
 		p := schemas.NewProject(env.ProjectName)
 
-		projectExists, err := c.Store.ProjectExists(p.Key())
+		projectExists, err := c.Store.ProjectExists(ctx, p.Key())
 		if err != nil {
-			log.WithFields(logFields).WithError(err).Error("reading project from the store")
+			log.WithContext(ctx).
+				WithFields(logFields).
+				WithError(err).
+				Error("reading project from the store")
+
 			return
 		}
 
 		// Perhaps the project is discoverable through a wildcard
 		if !projectExists && len(c.Config.Wildcards) > 0 {
-			for id, w := range c.Config.Wildcards {
+			for _, w := range c.Config.Wildcards {
 				// If in all our wildcards we have one which can potentially match the env
-				// received, we trigger a scan
+				// received, we trigger a pull of the project
 				matches, err := isEnvMatchingWilcard(w, env)
 				if err != nil {
-					log.WithError(err).Warn("checking if the env matches the wildcard config")
+					log.WithContext(ctx).
+						WithError(err).
+						Warn("checking if the env matches the wildcard config")
+
 					continue
 				}
 
 				if matches {
-					c.ScheduleTask(context.TODO(), schemas.TaskTypePullProjectsFromWildcard, strconv.Itoa(id), strconv.Itoa(id), w)
-					log.WithFields(logFields).Info("project environment not currently exported but its configuration matches a wildcard, triggering a pull of the projects from this wildcard")
+					c.ScheduleTask(context.TODO(), schemas.TaskTypePullProject, env.ProjectName, env.ProjectName, w.Pull)
+					log.WithFields(logFields).Info("project environment not currently exported but its configuration matches a wildcard, triggering a pull of the project")
 				} else {
 					log.WithFields(logFields).Debug("project ref not matching wildcard, skipping..")
 				}
 			}
+
 			log.WithFields(logFields).Info("done looking up for wildcards matching the project ref")
+
 			return
 		}
 
 		if projectExists {
-			if err := c.Store.GetProject(&p); err != nil {
-				log.WithFields(logFields).WithError(err).Error("reading project from the store")
+			if err := c.Store.GetProject(ctx, &p); err != nil {
+				log.WithContext(ctx).
+					WithFields(logFields).
+					WithError(err).
+					Error("reading project from the store")
 			}
 
 			matches, err := isEnvMatchingProjectPullEnvironments(p.Pull.Environments, env)
 			if err != nil {
-				log.WithError(err).Error("checking if the env matches the project config")
+				log.WithContext(ctx).
+					WithError(err).
+					Error("checking if the env matches the project config")
+
 				return
 			}
 
 			if matches {
 				// As we do not get the environment ID within the deployment event, we need to query it back..
-				if err = c.UpdateEnvironment(&env); err != nil {
-					log.WithFields(logFields).WithError(err).Error("updating event from GitLab API")
+				if err = c.UpdateEnvironment(ctx, &env); err != nil {
+					log.WithContext(ctx).
+						WithFields(logFields).
+						WithError(err).
+						Error("updating event from GitLab API")
+
 					return
 				}
+
 				goto schedulePull
 			}
 		}
 
-		log.WithFields(logFields).Info("environment not configured in the exporter, ignoring deployment webhook")
+		log.WithFields(logFields).
+			Info("environment not configured in the exporter, ignoring deployment webhook")
+
 		return
 	}
 
 	// Need to refresh the env from the store in order to get at least it's ID
 	if env.ID == 0 {
-		if err = c.Store.GetEnvironment(&env); err != nil {
-			log.WithFields(logFields).WithError(err).Error("reading environment from the store")
+		if err = c.Store.GetEnvironment(ctx, &env); err != nil {
+			log.WithContext(ctx).
+				WithFields(logFields).
+				WithError(err).
+				Error("reading environment from the store")
 		}
 	}
 
 schedulePull:
 	log.WithFields(logFields).Info("received a deployment webhook from GitLab for an environment, triggering metrics pull")
-	c.ScheduleTask(context.TODO(), schemas.TaskTypePullEnvironmentMetrics, string(env.Key()), env)
+	c.ScheduleTask(ctx, schemas.TaskTypePullEnvironmentMetrics, string(env.Key()), env)
 }
 
 func isRefMatchingProjectPullRefs(pprs config.ProjectPullRefs, ref schemas.Ref) (matches bool, err error) {
@@ -222,9 +389,11 @@ func isRefMatchingProjectPullRefs(pprs config.ProjectPullRefs, ref schemas.Ref) 
 
 	// Then we check if it matches the regexp
 	var re *regexp.Regexp
+
 	if re, err = schemas.GetRefRegexp(pprs, ref.Kind); err != nil {
 		return
 	}
+
 	return re.MatchString(ref.Name), nil
 }
 
@@ -236,9 +405,11 @@ func isEnvMatchingProjectPullEnvironments(ppe config.ProjectPullEnvironments, en
 
 	// Then we check if it matches the regexp
 	var re *regexp.Regexp
+
 	if re, err = regexp.Compile(ppe.Regexp); err != nil {
 		return
 	}
+
 	return re.MatchString(env.Name), nil
 }
 
