@@ -1,20 +1,22 @@
 package gitlab
 
 import (
+	"context"
 	"fmt"
 	"net/http"
 	"net/http/httptest"
 	"testing"
 	"time"
 
-	"github.com/mvisonneau/gitlab-ci-pipelines-exporter/pkg/ratelimit"
 	"github.com/paulbellamy/ratecounter"
 	"github.com/stretchr/testify/assert"
 	goGitlab "github.com/xanzy/go-gitlab"
+
+	"github.com/mvisonneau/gitlab-ci-pipelines-exporter/pkg/ratelimit"
 )
 
-// Mocking helpers
-func getMockedClient() (*http.ServeMux, *httptest.Server, *Client) {
+// Mocking helpers.
+func getMockedClient() (context.Context, *http.ServeMux, *httptest.Server, *Client) {
 	mux := http.NewServeMux()
 	server := httptest.NewServer(mux)
 
@@ -27,11 +29,11 @@ func getMockedClient() (*http.ServeMux, *httptest.Server, *Client) {
 
 	c := &Client{
 		Client:      gc,
-		RateLimiter: ratelimit.NewLocalLimiter(100),
+		RateLimiter: ratelimit.NewLocalLimiter(100, 1),
 		RateCounter: ratecounter.NewRateCounter(time.Second),
 	}
 
-	return mux, server, c
+	return context.Background(), mux, server, c
 }
 
 func TestNewHTTPClient(t *testing.T) {
@@ -44,9 +46,9 @@ func TestNewClient(t *testing.T) {
 		URL:              "https://gitlab.example.com",
 		Token:            "supersecret",
 		UserAgentVersion: "0.0.0",
-		DisableTLSVerify: false,
+		DisableTLSVerify: true,
 		ReadinessURL:     "https://gitlab.example.com/amialive",
-		RateLimiter:      ratelimit.NewLocalLimiter(10),
+		RateLimiter:      ratelimit.NewLocalLimiter(10, 1),
 	}
 
 	c, err := NewClient(cfg)
@@ -56,28 +58,35 @@ func TestNewClient(t *testing.T) {
 	assert.Equal(t, "https", c.Client.BaseURL().Scheme)
 	assert.Equal(t, "gitlab.example.com", c.Client.BaseURL().Host)
 	assert.Equal(t, "https://gitlab.example.com/amialive", c.Readiness.URL)
+	assert.True(t, c.Readiness.HTTPClient.Transport.(*http.Transport).TLSClientConfig.InsecureSkipVerify)
 	assert.Equal(t, 5*time.Second, c.Readiness.HTTPClient.Timeout)
 }
 
 func TestReadinessCheck(t *testing.T) {
-	mux, server, c := getMockedClient()
-	mux.HandleFunc(fmt.Sprintf("/200"),
+	ctx, mux, server, c := getMockedClient()
+	mux.HandleFunc(
+		"/200",
 		func(w http.ResponseWriter, r *http.Request) {
 			assert.Equal(t, "GET", r.Method)
 			w.WriteHeader(http.StatusOK)
-		})
-	mux.HandleFunc(fmt.Sprintf("/500"),
+		},
+	)
+	mux.HandleFunc(
+		"/500",
 		func(w http.ResponseWriter, r *http.Request) {
 			w.WriteHeader(http.StatusInternalServerError)
-		})
+		},
+	)
 
-	readinessCheck := c.ReadinessCheck()
+	readinessCheck := c.ReadinessCheck(ctx)
 	assert.Error(t, readinessCheck())
 
 	c.Readiness.HTTPClient = NewHTTPClient(false)
 	c.Readiness.URL = fmt.Sprintf("%s/200", server.URL)
+
 	assert.NoError(t, readinessCheck())
 
 	c.Readiness.URL = fmt.Sprintf("%s/500", server.URL)
+
 	assert.Error(t, readinessCheck())
 }
