@@ -1,18 +1,26 @@
 package gitlab
 
 import (
+	"context"
 	"regexp"
 
-	"github.com/mvisonneau/gitlab-ci-pipelines-exporter/pkg/schemas"
 	log "github.com/sirupsen/logrus"
 	goGitlab "github.com/xanzy/go-gitlab"
+	"go.opentelemetry.io/otel"
+	"go.opentelemetry.io/otel/attribute"
+
+	"github.com/mvisonneau/gitlab-ci-pipelines-exporter/pkg/schemas"
 )
 
 // GetProjectEnvironments ..
-func (c *Client) GetProjectEnvironments(p schemas.Project) (
+func (c *Client) GetProjectEnvironments(ctx context.Context, p schemas.Project) (
 	envs schemas.Environments,
 	err error,
 ) {
+	ctx, span := otel.Tracer(tracerName).Start(ctx, "gitlab:GetProjectEnvironments")
+	defer span.End()
+	span.SetAttributes(attribute.String("project_name", p.Name))
+
 	envs = make(schemas.Environments)
 
 	options := &goGitlab.ListEnvironmentsOptions{
@@ -32,13 +40,18 @@ func (c *Client) GetProjectEnvironments(p schemas.Project) (
 	}
 
 	for {
-		c.rateLimit()
-		var glenvs []*goGitlab.Environment
-		var resp *goGitlab.Response
-		glenvs, resp, err = c.Environments.ListEnvironments(p.Name, options)
+		c.rateLimit(ctx)
+
+		var (
+			glenvs []*goGitlab.Environment
+			resp   *goGitlab.Response
+		)
+
+		glenvs, resp, err = c.Environments.ListEnvironments(p.Name, options, goGitlab.WithContext(ctx))
 		if err != nil {
 			return
 		}
+
 		c.requestsRemaining(resp)
 
 		for _, glenv := range glenvs {
@@ -61,6 +74,7 @@ func (c *Client) GetProjectEnvironments(p schemas.Project) (
 		if resp.CurrentPage >= resp.NextPage {
 			break
 		}
+
 		options.Page = resp.NextPage
 	}
 
@@ -68,17 +82,36 @@ func (c *Client) GetProjectEnvironments(p schemas.Project) (
 }
 
 // GetEnvironment ..
-func (c *Client) GetEnvironment(project string, environmentID int) (schemas.Environment, error) {
-	environment := schemas.Environment{
+func (c *Client) GetEnvironment(
+	ctx context.Context,
+	project string,
+	environmentID int,
+) (
+	environment schemas.Environment,
+	err error,
+) {
+	ctx, span := otel.Tracer(tracerName).Start(ctx, "gitlab:GetEnvironment")
+	defer span.End()
+	span.SetAttributes(attribute.String("project_name", project))
+	span.SetAttributes(attribute.Int("environment_id", environmentID))
+
+	environment = schemas.Environment{
 		ProjectName: project,
 		ID:          environmentID,
 	}
 
-	c.rateLimit()
-	e, resp, err := c.Environments.GetEnvironment(project, environmentID, nil)
+	c.rateLimit(ctx)
+
+	var (
+		e    *goGitlab.Environment
+		resp *goGitlab.Response
+	)
+
+	e, resp, err = c.Environments.GetEnvironment(project, environmentID, goGitlab.WithContext(ctx))
 	if err != nil || e == nil {
-		return environment, err
+		return
 	}
+
 	c.requestsRemaining(resp)
 
 	environment.Name = e.Name
@@ -88,35 +121,39 @@ func (c *Client) GetEnvironment(project string, environmentID int) (schemas.Envi
 		environment.Available = true
 	}
 
-	if e.LastDeployment != nil {
-		if e.LastDeployment.Deployable.Tag {
-			environment.LatestDeployment.RefKind = schemas.RefKindTag
-		} else {
-			environment.LatestDeployment.RefKind = schemas.RefKindBranch
-		}
+	if e.LastDeployment == nil {
+		log.WithContext(ctx).
+			WithFields(log.Fields{
+				"project-name":     project,
+				"environment-name": e.Name,
+			}).
+			Debug("no deployments found for the environment")
 
-		environment.LatestDeployment.RefName = e.LastDeployment.Ref
-		environment.LatestDeployment.JobID = e.LastDeployment.Deployable.ID
-		environment.LatestDeployment.DurationSeconds = e.LastDeployment.Deployable.Duration
-		environment.LatestDeployment.Status = e.LastDeployment.Deployable.Status
-
-		if e.LastDeployment.Deployable.User != nil {
-			environment.LatestDeployment.Username = e.LastDeployment.Deployable.User.Username
-		}
-
-		if e.LastDeployment.Deployable.Commit != nil {
-			environment.LatestDeployment.CommitShortID = e.LastDeployment.Deployable.Commit.ShortID
-		}
-
-		if e.LastDeployment.CreatedAt != nil {
-			environment.LatestDeployment.Timestamp = float64(e.LastDeployment.CreatedAt.Unix())
-		}
-	} else {
-		log.WithFields(log.Fields{
-			"project-name":     project,
-			"environment-name": e.Name,
-		}).Warn("no deployments found for the environment")
+		return
 	}
 
-	return environment, nil
+	if e.LastDeployment.Deployable.Tag {
+		environment.LatestDeployment.RefKind = schemas.RefKindTag
+	} else {
+		environment.LatestDeployment.RefKind = schemas.RefKindBranch
+	}
+
+	environment.LatestDeployment.RefName = e.LastDeployment.Ref
+	environment.LatestDeployment.JobID = e.LastDeployment.Deployable.ID
+	environment.LatestDeployment.DurationSeconds = e.LastDeployment.Deployable.Duration
+	environment.LatestDeployment.Status = e.LastDeployment.Deployable.Status
+
+	if e.LastDeployment.Deployable.User != nil {
+		environment.LatestDeployment.Username = e.LastDeployment.Deployable.User.Username
+	}
+
+	if e.LastDeployment.Deployable.Commit != nil {
+		environment.LatestDeployment.CommitShortID = e.LastDeployment.Deployable.Commit.ShortID
+	}
+
+	if e.LastDeployment.CreatedAt != nil {
+		environment.LatestDeployment.Timestamp = float64(e.LastDeployment.CreatedAt.Unix())
+	}
+
+	return
 }
