@@ -5,29 +5,30 @@ import (
 	"net/http"
 	"testing"
 
-	"github.com/mvisonneau/gitlab-ci-pipelines-exporter/pkg/config"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+
+	"github.com/mvisonneau/gitlab-ci-pipelines-exporter/pkg/config"
 )
 
 func TestGetProject(t *testing.T) {
-	mux, server, c := getMockedClient()
+	ctx, mux, server, c := getMockedClient()
 	defer server.Close()
 
-	project := "foo/bar"
-	mux.HandleFunc(fmt.Sprintf("/api/v4/projects/%s", project),
+	mux.HandleFunc("/api/v4/projects/foo%2Fbar",
 		func(w http.ResponseWriter, r *http.Request) {
 			assert.Equal(t, r.Method, "GET")
-			fmt.Fprint(w, `{"id":1}`)
+			_, _ = fmt.Fprint(w, `{"id":1}`)
 		})
 
-	p, err := c.GetProject(project)
+	p, err := c.GetProject(ctx, "foo/bar")
 	assert.NoError(t, err)
-	assert.NotNil(t, p)
+	require.NotNil(t, p)
 	assert.Equal(t, 1, p.ID)
 }
 
 func TestListUserProjects(t *testing.T) {
-	mux, server, c := getMockedClient()
+	ctx, mux, server, c := getMockedClient()
 	defer server.Close()
 
 	w := config.Wildcard{
@@ -43,17 +44,17 @@ func TestListUserProjects(t *testing.T) {
 	mux.HandleFunc(fmt.Sprintf("/api/v4/users/%s/projects", w.Owner.Name),
 		func(w http.ResponseWriter, r *http.Request) {
 			assert.Equal(t, r.Method, "GET")
-			fmt.Fprint(w, `[{"id":1,"path_with_namespace":"foo/bar","jobs_enabled":true},{"id":2,"path_with_namespace":"bar/baz","jobs_enabled":true}]`)
+			_, _ = fmt.Fprint(w, `[{"id":1,"path_with_namespace":"foo/bar"},{"id":2,"path_with_namespace":"bar/baz"}]`)
 		})
 
-	projects, err := c.ListProjects(w)
+	projects, err := c.ListProjects(ctx, w)
 	assert.NoError(t, err)
 	assert.Len(t, projects, 1)
 	assert.Equal(t, "foo/bar", projects[0].Name)
 }
 
 func TestListGroupProjects(t *testing.T) {
-	mux, server, c := getMockedClient()
+	ctx, mux, server, c := getMockedClient()
 	defer server.Close()
 
 	w := config.Wildcard{
@@ -69,17 +70,17 @@ func TestListGroupProjects(t *testing.T) {
 	mux.HandleFunc(fmt.Sprintf("/api/v4/groups/%s/projects", w.Owner.Name),
 		func(w http.ResponseWriter, r *http.Request) {
 			assert.Equal(t, r.Method, "GET")
-			fmt.Fprint(w, `[{"id":1,"path_with_namespace":"foo/bar","jobs_enabled":true},{"id":2,"path_with_namespace":"bar/baz","jobs_enabled":true}]`)
+			_, _ = fmt.Fprint(w, `[{"id":1,"path_with_namespace":"foo/bar"},{"id":2,"path_with_namespace":"bar/baz"}]`)
 		})
 
-	projects, err := c.ListProjects(w)
+	projects, err := c.ListProjects(ctx, w)
 	assert.NoError(t, err)
 	assert.Len(t, projects, 1)
 	assert.Equal(t, "foo/bar", projects[0].Name)
 }
 
 func TestListProjects(t *testing.T) {
-	mux, server, c := getMockedClient()
+	ctx, mux, server, c := getMockedClient()
 	defer server.Close()
 
 	w := config.Wildcard{
@@ -95,17 +96,17 @@ func TestListProjects(t *testing.T) {
 	mux.HandleFunc("/api/v4/projects",
 		func(w http.ResponseWriter, r *http.Request) {
 			assert.Equal(t, r.Method, "GET")
-			fmt.Fprint(w, `[{"id":1,"path_with_namespace":"foo","jobs_enabled":false},{"id":2,"path_with_namespace":"bar","jobs_enabled":true}]`)
+			_, _ = fmt.Fprint(w, `[{"id":2,"path_with_namespace":"bar"}]`)
 		})
 
-	projects, err := c.ListProjects(w)
+	projects, err := c.ListProjects(ctx, w)
 	assert.NoError(t, err)
 	assert.Len(t, projects, 1)
 	assert.Equal(t, "bar", projects[0].Name)
 }
 
 func TestListProjectsAPIError(t *testing.T) {
-	mux, server, c := getMockedClient()
+	ctx, mux, server, c := getMockedClient()
 	defer server.Close()
 
 	w := config.Wildcard{
@@ -120,10 +121,10 @@ func TestListProjectsAPIError(t *testing.T) {
 	mux.HandleFunc(fmt.Sprintf("/api/v4/users/%s/projects", w.Owner.Name),
 		func(w http.ResponseWriter, r *http.Request) {
 			w.WriteHeader(http.StatusInternalServerError)
-			w.Write([]byte("500 - Something bad happened!"))
+			_, _ = w.Write([]byte("500 - Something bad happened!"))
 		})
 
-	_, err := c.ListProjects(w)
+	_, err := c.ListProjects(ctx, w)
 	assert.Error(t, err)
 	assert.Contains(t, err.Error(), "unable to list projects with search pattern")
 }
