@@ -1,17 +1,25 @@
 package gitlab
 
 import (
+	"context"
 	"regexp"
 
-	"github.com/mvisonneau/gitlab-ci-pipelines-exporter/pkg/schemas"
 	goGitlab "github.com/xanzy/go-gitlab"
+	"go.opentelemetry.io/otel"
+	"go.opentelemetry.io/otel/attribute"
+
+	"github.com/mvisonneau/gitlab-ci-pipelines-exporter/pkg/schemas"
 )
 
 // GetProjectTags ..
-func (c *Client) GetProjectTags(p schemas.Project) (
+func (c *Client) GetProjectTags(ctx context.Context, p schemas.Project) (
 	refs schemas.Refs,
 	err error,
 ) {
+	ctx, span := otel.Tracer(tracerName).Start(ctx, "gitlab:GetProjectTags")
+	defer span.End()
+	span.SetAttributes(attribute.String("project_name", p.Name))
+
 	refs = make(schemas.Refs)
 
 	options := &goGitlab.ListTagsOptions{
@@ -22,18 +30,24 @@ func (c *Client) GetProjectTags(p schemas.Project) (
 	}
 
 	var re *regexp.Regexp
+
 	if re, err = regexp.Compile(p.Pull.Refs.Tags.Regexp); err != nil {
 		return
 	}
 
 	for {
-		c.rateLimit()
-		var tags []*goGitlab.Tag
-		var resp *goGitlab.Response
-		tags, resp, err = c.Tags.ListTags(p.Name, options)
+		c.rateLimit(ctx)
+
+		var (
+			tags []*goGitlab.Tag
+			resp *goGitlab.Response
+		)
+
+		tags, resp, err = c.Tags.ListTags(p.Name, options, goGitlab.WithContext(ctx))
 		if err != nil {
 			return
 		}
+
 		c.requestsRemaining(resp)
 
 		for _, tag := range tags {
@@ -46,6 +60,7 @@ func (c *Client) GetProjectTags(p schemas.Project) (
 		if resp.CurrentPage >= resp.NextPage {
 			break
 		}
+
 		options.Page = resp.NextPage
 	}
 
@@ -53,7 +68,12 @@ func (c *Client) GetProjectTags(p schemas.Project) (
 }
 
 // GetProjectMostRecentTagCommit ..
-func (c *Client) GetProjectMostRecentTagCommit(projectName, filterRegexp string) (string, float64, error) {
+func (c *Client) GetProjectMostRecentTagCommit(ctx context.Context, projectName, filterRegexp string) (string, float64, error) {
+	ctx, span := otel.Tracer(tracerName).Start(ctx, "gitlab:GetProjectTags")
+	defer span.End()
+	span.SetAttributes(attribute.String("project_name", projectName))
+	span.SetAttributes(attribute.String("regexp", filterRegexp))
+
 	options := &goGitlab.ListTagsOptions{
 		ListOptions: goGitlab.ListOptions{
 			Page:    1,
@@ -67,11 +87,13 @@ func (c *Client) GetProjectMostRecentTagCommit(projectName, filterRegexp string)
 	}
 
 	for {
-		c.rateLimit()
-		tags, resp, err := c.Tags.ListTags(projectName, options)
+		c.rateLimit(ctx)
+
+		tags, resp, err := c.Tags.ListTags(projectName, options, goGitlab.WithContext(ctx))
 		if err != nil {
 			return "", 0, err
 		}
+
 		c.requestsRemaining(resp)
 
 		for _, tag := range tags {
@@ -83,6 +105,7 @@ func (c *Client) GetProjectMostRecentTagCommit(projectName, filterRegexp string)
 		if resp.CurrentPage >= resp.NextPage {
 			break
 		}
+
 		options.Page = resp.NextPage
 	}
 
