@@ -1,8 +1,7 @@
 package ui
 
 import (
-	"bufio"
-	"bytes"
+	"context"
 	"fmt"
 	"net/url"
 	"os"
@@ -10,45 +9,32 @@ import (
 	"strings"
 	"time"
 
-	chromaQuick "github.com/alecthomas/chroma/quick"
 	"github.com/charmbracelet/bubbles/progress"
 	"github.com/charmbracelet/bubbles/viewport"
 	tea "github.com/charmbracelet/bubbletea"
 	"github.com/charmbracelet/lipgloss"
-	"github.com/muesli/termenv"
-	"github.com/mvisonneau/gitlab-ci-pipelines-exporter/pkg/monitor"
-	"github.com/mvisonneau/gitlab-ci-pipelines-exporter/pkg/monitor/rpc"
 	log "github.com/sirupsen/logrus"
 	"github.com/xeonx/timeago"
-)
 
-var (
-	color   = termenv.ColorProfile().Color
-	keyword = termenv.Style{}.Foreground(color("204")).Background(color("235")).Styled
-	help    = termenv.Style{}.Foreground(color("241")).Styled
+	"github.com/mvisonneau/gitlab-ci-pipelines-exporter/pkg/monitor/client"
+	pb "github.com/mvisonneau/gitlab-ci-pipelines-exporter/pkg/monitor/protobuf"
 )
 
 type tab string
 
 const (
-	tabStatus tab = "status"
-	tabConfig tab = "config"
+	tabTelemetry tab = "telemetry"
+	tabConfig    tab = "config"
 )
 
 var tabs = [...]tab{
-	tabStatus,
+	tabTelemetry,
 	tabConfig,
 }
 
 var (
 	subtle    = lipgloss.AdaptiveColor{Light: "#D9DCCF", Dark: "#383838"}
 	highlight = lipgloss.AdaptiveColor{Light: "#874BFD", Dark: "#7D56F4"}
-	special   = lipgloss.AdaptiveColor{Light: "#43BF6D", Dark: "#73F59F"}
-	divider   = lipgloss.NewStyle().
-			SetString("•").
-			Padding(0, 1).
-			Foreground(subtle).
-			String()
 
 	dataStyle = lipgloss.NewStyle().
 			MarginLeft(1).
@@ -58,7 +44,7 @@ var (
 			Foreground(lipgloss.Color("#000000")).
 			Background(lipgloss.Color("#a9a9a9"))
 
-	// Tabs
+	// Tabs.
 
 	activeTabBorder = lipgloss.Border{
 		Top:         "─",
@@ -94,13 +80,13 @@ var (
 		BorderLeft(false).
 		BorderRight(false)
 
-	// List
+	// List.
 
 	entityStyle = lipgloss.NewStyle().
 			Border(lipgloss.NormalBorder(), true, false, false, false).
 			BorderForeground(subtle)
 
-	// Status Bar
+	// Status Bar.
 
 	statusStyle = lipgloss.NewStyle().
 			Inherit(statusBarStyle).
@@ -122,72 +108,73 @@ var (
 	versionStyle = statusNugget.Copy().
 			Background(lipgloss.Color("#0062cc"))
 
-	// Page
+	// Page.
 	docStyle = lipgloss.NewStyle()
 )
 
-func max(a, b int) int {
-	if a > b {
-		return a
-	}
-	return b
-}
-
 type model struct {
 	version         string
-	listenerAddress *url.URL
-	rpcClient       *rpc.Client
-	sub             chan monitor.Status
-	lastStatus      *monitor.Status
+	client          *client.Client
 	vp              viewport.Model
 	progress        *progress.Model
+	telemetry       *pb.Telemetry
+	telemetryStream chan *pb.Telemetry
 	tabID           int
 }
 
-func (m *model) renderLastStatus() string {
-	if m.lastStatus == nil {
+func (m *model) renderConfigViewport() string {
+	config, err := m.client.GetConfig(context.TODO(), &pb.Empty{})
+	if err != nil || config == nil {
+		log.WithError(err).Fatal()
+	}
+
+	return config.GetContent()
+}
+
+func (m *model) renderTelemetryViewport() string {
+	if m.telemetry == nil {
 		return "\nloading data.."
 	}
 
 	gitlabAPIUsage := lipgloss.JoinHorizontal(
 		lipgloss.Top,
 		" GitLab API usage        ",
-		m.progress.ViewAs(m.lastStatus.GitLabAPIUsage),
+		m.progress.ViewAs(m.telemetry.GitlabApiUsage),
 		"\n",
 	)
 
 	gitlabAPIRequestsCount := lipgloss.JoinHorizontal(
 		lipgloss.Top,
 		" GitLab API requests    ",
-		dataStyle.SetString(strconv.Itoa(int(m.lastStatus.GitLabAPIRequestsCount))).String(),
+		dataStyle.SetString(strconv.Itoa(int(m.telemetry.GetGitlabApiRequestsCount()))).String(),
 		"\n",
 	)
 
 	gitlabAPIRateLimit := lipgloss.JoinHorizontal(
 		lipgloss.Top,
 		" GitLab API limit usage  ",
-		m.progress.ViewAs(m.lastStatus.GitLabAPIRateLimit),
+		m.progress.ViewAs(m.telemetry.GetGitlabApiRateLimit()),
 		"\n",
 	)
 
 	gitlabAPIRateLimitRemaining := lipgloss.JoinHorizontal(
 		lipgloss.Top,
 		" GitLab API limit requests remaining ",
-		dataStyle.SetString(strconv.Itoa(int(m.lastStatus.GitLabAPILimitRemaining))).String(),
+		dataStyle.SetString(strconv.Itoa(int(m.telemetry.GetGitlabApiLimitRemaining()))).String(),
 		"\n",
 	)
 
 	tasksBufferUsage := lipgloss.JoinHorizontal(
 		lipgloss.Top,
 		" Tasks buffer usage      ",
-		m.progress.ViewAs(m.lastStatus.TasksBufferUsage),
+		m.progress.ViewAs(m.telemetry.GetTasksBufferUsage()),
 		"\n",
 	)
 
 	tasksExecuted := lipgloss.JoinHorizontal(
 		lipgloss.Top,
 		" Tasks executed         ",
-		dataStyle.SetString(strconv.Itoa(int(m.lastStatus.TasksExecutedCount))).String(),
+		dataStyle.SetString(strconv.Itoa(int(m.telemetry.GetTasksExecutedCount()))).String(),
 		"\n",
 	)
 
@@ -199,24 +186,24 @@ func (m *model) renderLastStatus() string {
 		gitlabAPIRateLimitRemaining,
 		tasksBufferUsage,
 		tasksExecuted,
-		renderEntityStatus("Projects", m.lastStatus.Projects),
-		renderEntityStatus("Environments", m.lastStatus.Envs),
-		renderEntityStatus("Refs", m.lastStatus.Refs),
-		renderEntityStatus("Metrics", m.lastStatus.Metrics),
+		renderEntity("Projects", m.telemetry.GetProjects()),
+		renderEntity("Environments", m.telemetry.GetEnvs()),
+		renderEntity("Refs", m.telemetry.GetRefs()),
+		renderEntity("Metrics", m.telemetry.GetMetrics()),
 	}, "\n")
 }
 
-func renderEntityStatus(name string, es monitor.EntityStatus) string {
+func renderEntity(name string, e *pb.Entity) string {
 	return entityStyle.Render(lipgloss.JoinHorizontal(
 		lipgloss.Top,
 		" "+name+strings.Repeat(" ", 24-len(name)),
 		lipgloss.JoinVertical(
 			lipgloss.Left,
-			"Total      "+dataStyle.SetString(strconv.Itoa(int(es.Count))).String()+"\n",
-			"Last Pull  "+dataStyle.SetString(prettyTimeago(es.LastPull)).String()+"\n",
-			"Last GC    "+dataStyle.SetString(prettyTimeago(es.LastGC)).String()+"\n",
-			"Next Pull  "+dataStyle.SetString(prettyTimeago(es.NextPull)).String()+"\n",
-			"Next GC    "+dataStyle.SetString(prettyTimeago(es.NextGC)).String()+"\n",
+			"Total      "+dataStyle.SetString(strconv.Itoa(int(e.Count))).String()+"\n",
+			"Last Pull  "+dataStyle.SetString(prettyTimeago(e.LastPull.AsTime())).String()+"\n",
+			"Last GC    "+dataStyle.SetString(prettyTimeago(e.LastGc.AsTime())).String()+"\n",
+			"Next Pull  "+dataStyle.SetString(prettyTimeago(e.NextPull.AsTime())).String()+"\n",
+			"Next GC    "+dataStyle.SetString(prettyTimeago(e.NextGc.AsTime())).String()+"\n",
 		),
 		"\n",
 	))
@@ -226,38 +213,39 @@ func prettyTimeago(t time.Time) string {
 	if t.IsZero() {
 		return "N/A"
 	}
+
 	return timeago.English.Format(t)
 }
 
-func newModel(version string, listenerAddress *url.URL) (m *model) {
-	rpcClient := rpc.NewClient(listenerAddress)
+func newModel(version string, endpoint *url.URL) (m *model) {
 	p := progress.NewModel(progress.WithScaledGradient("#80c904", "#ff9d5c"))
 
 	m = &model{
 		version:         version,
-		listenerAddress: listenerAddress,
-		sub:             make(chan monitor.Status),
 		vp:              viewport.Model{},
+		telemetryStream: make(chan *pb.Telemetry),
 		progress:        &p,
-		rpcClient:       rpcClient,
+		client:          client.NewClient(context.TODO(), endpoint),
 	}
+
 	return
 }
 
-func (m model) Init() tea.Cmd {
+func (m *model) Init() tea.Cmd {
 	return tea.Batch(
-		m.generateActivity(),
-		waitForActivity(m.sub),
+		m.streamTelemetry(context.TODO()),
+		waitForTelemetryUpdate(m.telemetryStream),
 	)
 }
 
-func (m model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
+func (m *model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 	switch msg := msg.(type) {
 	case tea.WindowSizeMsg:
 		m.vp.Width = msg.Width
 		m.vp.Height = msg.Height - 4
 		m.progress.Width = msg.Width - 27
 		m.setPaneContent()
+
 		return m, nil
 	case tea.KeyMsg:
 		switch msg.Type {
@@ -268,40 +256,45 @@ func (m model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 				m.tabID--
 				m.setPaneContent()
 			}
+
 			return m, nil
 		case tea.KeyRight:
 			if m.tabID < len(tabs)-1 {
 				m.tabID++
 				m.setPaneContent()
 			}
+
 			return m, nil
 		case tea.KeyUp, tea.KeyDown, tea.KeyPgDown, tea.KeyPgUp:
 			vp, cmd := m.vp.Update(msg)
 			m.vp = vp
+
 			return m, cmd
 		}
-	case monitor.Status:
-		m.lastStatus = &msg
-		if m.tabID == 0 {
-			m.vp.SetContent(m.renderLastStatus())
-		}
-		return m, waitForActivity(m.sub)
+	case *pb.Telemetry:
+		m.telemetry = msg
+		m.setPaneContent()
+
+		return m, waitForTelemetryUpdate(m.telemetryStream)
 	}
 
 	return m, nil
 }
 
-func (m model) View() string {
+func (m *model) View() string {
 	doc := strings.Builder{}
 
 	// TABS
 	{
 		renderedTabs := []string{}
+
 		for tabID, t := range tabs {
 			if m.tabID == tabID {
 				renderedTabs = append(renderedTabs, activeTab.Render(string(t)))
+
 				continue
 			}
+
 			renderedTabs = append(renderedTabs, inactiveTab.Render(string(t)))
 		}
 
@@ -311,12 +304,12 @@ func (m model) View() string {
 		doc.WriteString(row + "\n")
 	}
 
-	// PANE
+	// Pane.
 	{
 		doc.WriteString(m.vp.View() + "\n")
 	}
 
-	// Status bar
+	// Status bar.
 	{
 		bar := lipgloss.JoinHorizontal(lipgloss.Top,
 			statusStyle.Render("github.com/mvisonneau/gitlab-ci-pipelines-exporter"),
@@ -332,18 +325,29 @@ func (m model) View() string {
 	return docStyle.Render(doc.String())
 }
 
-func waitForActivity(sub chan monitor.Status) tea.Cmd {
-	return func() tea.Msg {
-		return <-sub
+func (m *model) streamTelemetry(ctx context.Context) tea.Cmd {
+	c, err := m.client.GetTelemetry(ctx, &pb.Empty{})
+	if err != nil {
+		log.WithError(err).Fatal()
 	}
+
+	go func(m *model) {
+		for {
+			telemetry, err := c.Recv()
+			if err != nil {
+				log.WithError(err).Fatal()
+			}
+
+			m.telemetryStream <- telemetry
+		}
+	}(m)
+
+	return nil
 }
 
-func (m model) generateActivity() tea.Cmd {
+func waitForTelemetryUpdate(t chan *pb.Telemetry) tea.Cmd {
 	return func() tea.Msg {
-		for {
-			time.Sleep(time.Second)
-			m.sub <- m.rpcClient.Status()
-		}
+		return <-t
 	}
 }
 
@@ -360,15 +364,9 @@ func Start(version string, listenerAddress *url.URL) {
 
 func (m *model) setPaneContent() {
 	switch tabs[m.tabID] {
-	case tabStatus:
-		m.vp.SetContent(m.renderLastStatus())
+	case tabTelemetry:
+		m.vp.SetContent(m.renderTelemetryViewport())
 	case tabConfig:
-		var b bytes.Buffer
-		foo := bufio.NewWriter(&b)
-		if err := chromaQuick.Highlight(foo, m.rpcClient.Config(), "yaml", "terminal16m", "monokai"); err != nil {
-			log.WithError(err).Fatal()
-		}
-
-		m.vp.SetContent(b.String())
+		m.vp.SetContent(m.renderConfigViewport())
 	}
 }
