@@ -1,26 +1,16 @@
 package ratelimit
 
 import (
+	"context"
 	"time"
-
-	log "github.com/sirupsen/logrus"
 )
 
 // Limiter ..
 type Limiter interface {
-	Take() time.Time
+	Take(ctx context.Context) time.Duration
 }
 
 // Take ..
-func Take(l Limiter) {
-	now := time.Now()
-	throttled := l.Take()
-	if throttled.Sub(now).Milliseconds() > 10 {
-		log.WithFields(
-			log.Fields{
-				"for": throttled.Sub(now),
-			},
-		).Debug("throttled GitLab requests")
-	}
-	return
+func Take(ctx context.Context, l Limiter) {
+	l.Take(ctx)
 }
