@@ -4,8 +4,8 @@ import (
 	"context"
 	"time"
 
-	"github.com/go-redis/redis/v8"
-	"github.com/go-redis/redis_rate/v9"
+	"github.com/go-redis/redis_rate/v10"
+	"github.com/redis/go-redis/v9"
 	log "github.com/sirupsen/logrus"
 )
 
@@ -14,25 +14,40 @@ const redisKey string = `gcpe:gitlab:api`
 // Redis ..
 type Redis struct {
 	*redis_rate.Limiter
-	Context context.Context
-	MaxRPS  int
+	MaxRPS int
 }
 
 // NewRedisLimiter ..
-func NewRedisLimiter(ctx context.Context, redisClient *redis.Client, maxRPS int) Limiter {
+func NewRedisLimiter(redisClient *redis.Client, maxRPS int) Limiter {
 	return Redis{
 		Limiter: redis_rate.NewLimiter(redisClient),
-		Context: ctx,
 		MaxRPS:  maxRPS,
 	}
 }
 
 // Take ..
-func (r Redis) Take() time.Time {
-	res, err := r.Allow(r.Context, redisKey, redis_rate.PerSecond(r.MaxRPS))
-	if err != nil {
-		log.Fatalf(err.Error())
+func (r Redis) Take(ctx context.Context) time.Duration {
+	start := time.Now()
+
+	for {
+		res, err := r.Allow(ctx, redisKey, redis_rate.PerSecond(r.MaxRPS))
+		if err != nil {
+			log.WithContext(ctx).
+				WithError(err).
+				Fatal()
+		}
+
+		if res.Allowed > 0 {
+			break
+		} else {
+			log.WithFields(
+				log.Fields{
+					"for": res.RetryAfter.String(),
+				},
+			).Debug("throttled GitLab requests")
+			time.Sleep(res.RetryAfter)
+		}
 	}
-	time.Sleep(res.RetryAfter)
-	return time.Now()
+
+	return start.Sub(time.Now())
 }
