@@ -9,7 +9,7 @@ import (
 )
 
 const (
-	// MetricKindCoverage refers to the coerage of a job/pipeline
+	// MetricKindCoverage refers to the coerage of a job/pipeline.
 	MetricKindCoverage MetricKind = iota
 
 	// MetricKindDurationSeconds ..
@@ -74,6 +74,48 @@ const (
 
 	// MetricKindTimestamp ..
 	MetricKindTimestamp
+
+	// MetricKindTestReportTotalTime ..
+	MetricKindTestReportTotalTime
+
+	// MetricKindTestReportTotalCount ..
+	MetricKindTestReportTotalCount
+
+	// MetricKindTestReportSuccessCount ..
+	MetricKindTestReportSuccessCount
+
+	// MetricKindTestReportFailedCount ..
+	MetricKindTestReportFailedCount
+
+	// MetricKindTestReportSkippedCount ..
+	MetricKindTestReportSkippedCount
+
+	// MetricKindTestReportErrorCount ..
+	MetricKindTestReportErrorCount
+
+	// MetricKindTestSuiteTotalTime ..
+	MetricKindTestSuiteTotalTime
+
+	// MetricKindTestSuiteTotalCount ..
+	MetricKindTestSuiteTotalCount
+
+	// MetricKindTestSuiteSuccessCount ..
+	MetricKindTestSuiteSuccessCount
+
+	// MetricKindTestSuiteFailedCount ..
+	MetricKindTestSuiteFailedCount
+
+	// MetricKindTestSuiteSkippedCount ..
+	MetricKindTestSuiteSkippedCount
+
+	// MetricKindTestSuiteErrorCount ..
+	MetricKindTestSuiteErrorCount
+
+	// MetricKindTestCaseExecutionTime ..
+	MetricKindTestCaseExecutionTime
+
+	// MetricKindTestCaseStatus ..
+	MetricKindTestCaseStatus
 )
 
 // MetricKind ..
@@ -97,11 +139,12 @@ func (m Metric) Key() MetricKey {
 	key := strconv.Itoa(int(m.Kind))
 
 	switch m.Kind {
-	case MetricKindCoverage, MetricKindDurationSeconds, MetricKindID, MetricKindQueuedDurationSeconds, MetricKindRunCount, MetricKindStatus, MetricKindTimestamp:
+	case MetricKindCoverage, MetricKindDurationSeconds, MetricKindID, MetricKindQueuedDurationSeconds, MetricKindRunCount, MetricKindStatus, MetricKindTimestamp, MetricKindTestReportTotalCount, MetricKindTestReportErrorCount, MetricKindTestReportFailedCount, MetricKindTestReportSkippedCount, MetricKindTestReportSuccessCount, MetricKindTestReportTotalTime:
 		key += fmt.Sprintf("%v", []string{
 			m.Labels["project"],
 			m.Labels["kind"],
 			m.Labels["ref"],
+			m.Labels["source"],
 		})
 
 	case MetricKindJobArtifactSizeBytes, MetricKindJobDurationSeconds, MetricKindJobID, MetricKindJobQueuedDurationSeconds, MetricKindJobRunCount, MetricKindJobStatus, MetricKindJobTimestamp:
@@ -110,7 +153,9 @@ func (m Metric) Key() MetricKey {
 			m.Labels["kind"],
 			m.Labels["ref"],
 			m.Labels["stage"],
+			m.Labels["tag_list"],
 			m.Labels["job_name"],
+			m.Labels["failure_reason"],
 		})
 
 	case MetricKindEnvironmentBehindCommitsCount, MetricKindEnvironmentBehindDurationSeconds, MetricKindEnvironmentDeploymentCount, MetricKindEnvironmentDeploymentDurationSeconds, MetricKindEnvironmentDeploymentJobID, MetricKindEnvironmentDeploymentStatus, MetricKindEnvironmentDeploymentTimestamp, MetricKindEnvironmentInformation:
@@ -118,11 +163,29 @@ func (m Metric) Key() MetricKey {
 			m.Labels["project"],
 			m.Labels["environment"],
 		})
+
+	case MetricKindTestSuiteErrorCount, MetricKindTestSuiteFailedCount, MetricKindTestSuiteSkippedCount, MetricKindTestSuiteSuccessCount, MetricKindTestSuiteTotalCount, MetricKindTestSuiteTotalTime:
+		key += fmt.Sprintf("%v", []string{
+			m.Labels["project"],
+			m.Labels["kind"],
+			m.Labels["ref"],
+			m.Labels["test_suite_name"],
+		})
+
+	case MetricKindTestCaseExecutionTime, MetricKindTestCaseStatus:
+		key += fmt.Sprintf("%v", []string{
+			m.Labels["project"],
+			m.Labels["kind"],
+			m.Labels["ref"],
+			m.Labels["test_suite_name"],
+			m.Labels["test_case_name"],
+			m.Labels["test_case_classname"],
+		})
 	}
 
 	// If the metric is a "status" one, add the status label
 	switch m.Kind {
-	case MetricKindJobStatus, MetricKindEnvironmentDeploymentStatus, MetricKindStatus:
+	case MetricKindJobStatus, MetricKindEnvironmentDeploymentStatus, MetricKindStatus, MetricKindTestCaseStatus:
 		key += m.Labels["status"]
 	}
 
