@@ -3,48 +3,47 @@ package store
 import (
 	"context"
 
-	"github.com/go-redis/redis/v8"
+	"github.com/redis/go-redis/v9"
+	log "github.com/sirupsen/logrus"
+	"go.opentelemetry.io/otel"
+
 	"github.com/mvisonneau/gitlab-ci-pipelines-exporter/pkg/config"
 	"github.com/mvisonneau/gitlab-ci-pipelines-exporter/pkg/schemas"
-	log "github.com/sirupsen/logrus"
 )
 
 // Store ..
 type Store interface {
-	SetProject(schemas.Project) error
-	DelProject(schemas.ProjectKey) error
-	GetProject(*schemas.Project) error
-	ProjectExists(schemas.ProjectKey) (bool, error)
-	Projects() (schemas.Projects, error)
-	ProjectsCount() (int64, error)
-
-	SetEnvironment(schemas.Environment) error
-	DelEnvironment(schemas.EnvironmentKey) error
-	GetEnvironment(*schemas.Environment) error
-	EnvironmentExists(schemas.EnvironmentKey) (bool, error)
-	Environments() (schemas.Environments, error)
-	EnvironmentsCount() (int64, error)
-
-	SetRef(schemas.Ref) error
-	DelRef(schemas.RefKey) error
-	GetRef(*schemas.Ref) error
-	RefExists(schemas.RefKey) (bool, error)
-	Refs() (schemas.Refs, error)
-	RefsCount() (int64, error)
-
-	SetMetric(schemas.Metric) error
-	DelMetric(schemas.MetricKey) error
-	GetMetric(*schemas.Metric) error
-	MetricExists(schemas.MetricKey) (bool, error)
-	Metrics() (schemas.Metrics, error)
-	MetricsCount() (int64, error)
+	SetProject(ctx context.Context, p schemas.Project) error
+	DelProject(ctx context.Context, pk schemas.ProjectKey) error
+	GetProject(ctx context.Context, p *schemas.Project) error
+	ProjectExists(ctx context.Context, pk schemas.ProjectKey) (bool, error)
+	Projects(ctx context.Context) (schemas.Projects, error)
+	ProjectsCount(ctx context.Context) (int64, error)
+	SetEnvironment(ctx context.Context, e schemas.Environment) error
+	DelEnvironment(ctx context.Context, ek schemas.EnvironmentKey) error
+	GetEnvironment(ctx context.Context, e *schemas.Environment) error
+	EnvironmentExists(ctx context.Context, ek schemas.EnvironmentKey) (bool, error)
+	Environments(ctx context.Context) (schemas.Environments, error)
+	EnvironmentsCount(ctx context.Context) (int64, error)
+	SetRef(ctx context.Context, r schemas.Ref) error
+	DelRef(ctx context.Context, rk schemas.RefKey) error
+	GetRef(ctx context.Context, r *schemas.Ref) error
+	RefExists(ctx context.Context, rk schemas.RefKey) (bool, error)
+	Refs(ctx context.Context) (schemas.Refs, error)
+	RefsCount(ctx context.Context) (int64, error)
+	SetMetric(ctx context.Context, m schemas.Metric) error
+	DelMetric(ctx context.Context, mk schemas.MetricKey) error
+	GetMetric(ctx context.Context, m *schemas.Metric) error
+	MetricExists(ctx context.Context, mk schemas.MetricKey) (bool, error)
+	Metrics(ctx context.Context) (schemas.Metrics, error)
+	MetricsCount(ctx context.Context) (int64, error)
 
 	// Helpers to keep track of currently queued tasks and avoid scheduling them
 	// twice at the risk of ending up with loads of dangling goroutines being locked
-	QueueTask(schemas.TaskType, string, string) (bool, error)
-	UnqueueTask(schemas.TaskType, string) error
-	CurrentlyQueuedTasksCount() (uint64, error)
-	ExecutedTasksCount() (uint64, error)
+	QueueTask(ctx context.Context, tt schemas.TaskType, taskUUID, processUUID string) (bool, error)
+	UnqueueTask(ctx context.Context, tt schemas.TaskType, taskUUID string) error
+	CurrentlyQueuedTasksCount(ctx context.Context) (uint64, error)
+	ExecutedTasksCount(ctx context.Context) (uint64, error)
 }
 
 // NewLocalStore ..
@@ -61,16 +60,19 @@ func NewLocalStore() Store {
 func NewRedisStore(client *redis.Client) Store {
 	return &Redis{
 		Client: client,
-		ctx:    context.TODO(),
 	}
 }
 
 // New creates a new store and populates it with
-// provided []schemas.Project
+// provided []schemas.Project.
 func New(
+	ctx context.Context,
 	r *redis.Client,
 	projects config.Projects,
 ) (s Store) {
+	ctx, span := otel.Tracer("gitlab-ci-pipelines-exporter").Start(ctx, "store:New")
+	defer span.End()
+
 	if r != nil {
 		s = NewRedisStore(r)
 	} else {
@@ -80,20 +82,25 @@ func New(
 	// Load all the configured projects in the store
 	for _, p := range projects {
 		sp := schemas.Project{Project: p}
-		exists, err := s.ProjectExists(sp.Key())
+
+		exists, err := s.ProjectExists(ctx, sp.Key())
 		if err != nil {
-			log.WithFields(log.Fields{
-				"project-name": p.Name,
-				"error":        err.Error(),
-			}).Error("reading project from the store")
+			log.WithContext(ctx).
+				WithFields(log.Fields{
+					"project-name": p.Name,
+				}).
+				WithError(err).
+				Error("reading project from the store")
 		}
 
 		if !exists {
-			if err = s.SetProject(sp); err != nil {
-				log.WithFields(log.Fields{
-					"project-name": p.Name,
-					"error":        err.Error(),
-				}).Error("writing project in the store")
+			if err = s.SetProject(ctx, sp); err != nil {
+				log.WithContext(ctx).
+					WithFields(log.Fields{
+						"project-name": p.Name,
+					}).
+					WithError(err).
+					Error("writing project in the store")
 			}
 		}
 	}
