@@ -4,11 +4,14 @@ import (
 	"context"
 	"testing"
 
-	"github.com/go-redis/redis/v8"
+	"github.com/redis/go-redis/v9"
+	"github.com/stretchr/testify/assert"
+
 	"github.com/mvisonneau/gitlab-ci-pipelines-exporter/pkg/config"
 	"github.com/mvisonneau/gitlab-ci-pipelines-exporter/pkg/schemas"
-	"github.com/stretchr/testify/assert"
 )
+
+var testCtx = context.Background()
 
 func TestNewLocalStore(t *testing.T) {
 	expectedValue := &Local{
@@ -24,21 +27,20 @@ func TestNewRedisStore(t *testing.T) {
 	redisClient := redis.NewClient(&redis.Options{})
 	expectedValue := &Redis{
 		Client: redisClient,
-		ctx:    context.TODO(),
 	}
 
 	assert.Equal(t, expectedValue, NewRedisStore(redisClient))
 }
 
 func TestNew(t *testing.T) {
-	localStore := New(nil, config.Projects{})
+	localStore := New(testCtx, nil, config.Projects{})
 	assert.IsType(t, &Local{}, localStore)
 
 	redisClient := redis.NewClient(&redis.Options{})
-	redisStore := New(redisClient, config.Projects{})
+	redisStore := New(testCtx, redisClient, config.Projects{})
 	assert.IsType(t, &Redis{}, redisStore)
 
-	localStore = New(nil, config.Projects{
+	localStore = New(testCtx, nil, config.Projects{
 		{
 			Name: "foo",
 		},
@@ -49,6 +51,6 @@ func TestNew(t *testing.T) {
 			Name: "bar",
 		},
 	})
-	count, _ := localStore.ProjectsCount()
+	count, _ := localStore.ProjectsCount(testCtx)
 	assert.Equal(t, int64(2), count)
 }
